+++ lib/subdir.am
@@ -159,6 +159,7 @@ pkginclude_HEADERS += \
 	lib/graph.h \
 	lib/hash.h \
 	lib/hook.h \
+	lib/iana_afi.h \
 	lib/id_alloc.h \
 	lib/if.h \
 	lib/if_rmap.h \
