+++ lib/zebra.h
@@ -442,33 +442,6 @@ typedef enum {
 	SAFI_MAX = 8
 } safi_t;
 
-/*
- * The above AFI and SAFI definitions are for internal use. The protocol
- * definitions (IANA values) as for example used in BGP protocol packets
- * are defined below and these will get mapped to/from the internal values
- * in the appropriate places.
- * The rationale is that the protocol (IANA) values may be sparse and are
- * not optimal for use in data-structure sizing.
- * Note: Only useful (i.e., supported) values are defined below.
- */
-typedef enum {
-	IANA_AFI_RESERVED = 0,
-	IANA_AFI_IPV4 = 1,
-	IANA_AFI_IPV6 = 2,
-	IANA_AFI_L2VPN = 25,
-} iana_afi_t;
-
-typedef enum {
-	IANA_SAFI_RESERVED = 0,
-	IANA_SAFI_UNICAST = 1,
-	IANA_SAFI_MULTICAST = 2,
-	IANA_SAFI_LABELED_UNICAST = 4,
-	IANA_SAFI_ENCAP = 7,
-	IANA_SAFI_EVPN = 70,
-	IANA_SAFI_MPLS_VPN = 128,
-	IANA_SAFI_FLOWSPEC = 133
-} iana_safi_t;
-
 /* Default Administrative Distance of each protocol. */
 #define ZEBRA_KERNEL_DISTANCE_DEFAULT      0
 #define ZEBRA_CONNECT_DISTANCE_DEFAULT     0
@@ -509,117 +482,5 @@ typedef uint32_t vrf_id_t;
 typedef uint32_t route_tag_t;
 #define ROUTE_TAG_MAX UINT32_MAX
 #define ROUTE_TAG_PRI PRIu32
-
-static inline const char *iana_afi2str(iana_afi_t afi)
-{
-	switch (afi) {
-	case IANA_AFI_RESERVED:
-		return "Reserved";
-	case IANA_AFI_IPV4:
-		return "IPv4";
-	case IANA_AFI_IPV6:
-		return "IPv6";
-	case IANA_AFI_L2VPN:
-		return "L2VPN";
-	}
-
-	return "Unknown";
-}
-
-static inline afi_t afi_iana2int(iana_afi_t afi)
-{
-	switch (afi) {
-	case IANA_AFI_IPV4:
-		return AFI_IP;
-	case IANA_AFI_IPV6:
-		return AFI_IP6;
-	case IANA_AFI_L2VPN:
-		return AFI_L2VPN;
-	default:
-		return AFI_MAX;
-	}
-}
-
-static inline iana_afi_t afi_int2iana(afi_t afi)
-{
-	switch (afi) {
-	case AFI_IP:
-		return IANA_AFI_IPV4;
-	case AFI_IP6:
-		return IANA_AFI_IPV6;
-	case AFI_L2VPN:
-		return IANA_AFI_L2VPN;
-	default:
-		return IANA_AFI_RESERVED;
-	}
-}
-
-static inline const char *iana_safi2str(iana_safi_t safi)
-{
-	switch (safi) {
-	case IANA_SAFI_RESERVED:
-		return "Reserved";
-	case IANA_SAFI_UNICAST:
-		return "Unicast";
-	case IANA_SAFI_MULTICAST:
-		return "Multicast";
-	case IANA_SAFI_LABELED_UNICAST:
-		return "Labeled Unicast";
-	case IANA_SAFI_ENCAP:
-		return "Encap";
-	case IANA_SAFI_EVPN:
-		return "EVPN";
-	case IANA_SAFI_MPLS_VPN:
-		return "MPLS VPN";
-	case IANA_SAFI_FLOWSPEC:
-		return "FlowSpec";
-	}
-
-	return "Unknown";
-}
-
-static inline safi_t safi_iana2int(iana_safi_t safi)
-{
-	switch (safi) {
-	case IANA_SAFI_UNICAST:
-		return SAFI_UNICAST;
-	case IANA_SAFI_MULTICAST:
-		return SAFI_MULTICAST;
-	case IANA_SAFI_MPLS_VPN:
-		return SAFI_MPLS_VPN;
-	case IANA_SAFI_ENCAP:
-		return SAFI_ENCAP;
-	case IANA_SAFI_EVPN:
-		return SAFI_EVPN;
-	case IANA_SAFI_LABELED_UNICAST:
-		return SAFI_LABELED_UNICAST;
-	case IANA_SAFI_FLOWSPEC:
-		return SAFI_FLOWSPEC;
-	default:
-		return SAFI_MAX;
-	}
-}
-
-static inline iana_safi_t safi_int2iana(safi_t safi)
-{
-	switch (safi) {
-	case SAFI_UNICAST:
-		return IANA_SAFI_UNICAST;
-	case SAFI_MULTICAST:
-		return IANA_SAFI_MULTICAST;
-	case SAFI_MPLS_VPN:
-		return IANA_SAFI_MPLS_VPN;
-	case SAFI_ENCAP:
-		return IANA_SAFI_ENCAP;
-	case SAFI_EVPN:
-		return IANA_SAFI_EVPN;
-	case SAFI_LABELED_UNICAST:
-		return IANA_SAFI_LABELED_UNICAST;
-	case SAFI_FLOWSPEC:
-		return IANA_SAFI_FLOWSPEC;
-	default:
-		return IANA_SAFI_RESERVED;
-	}
-}
 
 #endif /* _ZEBRA_H */
